requirements.txt
beautifulsoup4==4.8.2
blis==0.4.1
Bottleneck==1.3.2
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cycler==0.10.0
cymem==2.0.3
fastai==1.0.60
fastapi==0.53.0
fastprogress==0.2.2
h11==0.9.0
idna==2.9
importlib-metadata==1.6.0
kiwisolver==1.1.0
matplotlib==3.2.1
murmurhash==1.0.2
numexpr==2.7.1
numpy==1.18.2
nvidia-ml-py3==7.352.0
packaging==20.3
pandas==1.0.3
Pillow==7.0.0
plac==1.1.3
preshed==3.0.2
pydantic==1.4
pyparsing==2.4.6
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
scipy==1.4.1
six==1.14.0
soupsieve==2.0
spacy==2.2.4
srsly==1.0.2
starlette==0.13.2
thinc==7.4.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.44.1
urllib3==1.25.8
uvicorn==0.11.3
wasabi==0.6.0
websockets==8.1
zipp==3.1.0
